feat: refactor duration
